refactor(check): colocate remaining tests into home modules; delete tests.rs monolith
Move all 128 tests out of the check/tests.rs monolith into the module each
exercises:
- check_completeness orchestrator tests + preserved_by/build_* cluster -> lints/mod.rs
- contains_word / parse_top_level_cmp -> lints/shared.rs
- CPI lints (multi_cpi, shape_only, unverified-callee, require-verified) -> lints/cpi.rs
- error-as-record / unknown_error_variant / pda / invariant_no_body / vpl -> lints/structural.rs
- old-in-single-state / cross-adt / unguarded / unconstrained_modifies -> lints/state.rs
- wrapping/saturating, ref_impl, map+subscript, missing_math_overflow -> lints/arithmetic.rs
- unbound_auth -> lints/auth.rs
- coverage_matrix -> check/coverage.rs
- parse_spec_file / multi-file / pragma sbpf / interface / ML syntax -> check/parse.rs

Shared builders empty_spec()/make_handler() centralized in a new
#[cfg(test)] pub(crate) mod test_support; synthetic_handler_default stays
local to lints/mod.rs. Section-banner comments stay attached to their tests.

Move-only: no assertion changes. tests.rs deleted; check/mod.rs no longer
declares `mod tests`. Unit suite 1006 passed / 0 failed; all snapshot suites
green.
