+++ crates/qedgen/src/check/lints/cpi.rs
@@ -415,3 +415,647 @@ pub(crate) fn check_shape_only_cpi(spec: &ParsedSpec) -> Vec<CompletenessWarning
 
     warnings
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    // ========================================================================
+    // multi_cpi_same_field lint
+    // ========================================================================
+
+    /// Two CPI calls whose substituted ensures both reference the same
+    /// caller-state field (`post.vault_balance`) → lint fires P2 Info.
+    /// Mirrors the bear-hug scenario where two `Token.transfer` calls
+    /// drain the same vault. Without per-call snapshot frames (v3.0),
+    /// the Kani harness can over-constrain.
+    #[test]
+    fn multi_cpi_same_field_fires_on_two_token_transfers_from_same_vault() {
+        let src = r#"spec MultiCpi
+    program_id "11111111111111111111111111111111"
+
+    interface Token {
+      program_id "11111111111111111111111111111111"
+      handler transfer (amount : U64) {
+        accounts {
+          from      : writable
+          to        : writable
+          authority : signer
+        }
+        requires amount > 0
+        ensures state.vault_balance == old(state.vault_balance) - amount
+      }
+    }
+
+    state { vault_balance : U64 }
+
+    handler split (a : U64) (b : U64) {
+      permissionless
+      requires a > 0 else InvalidAmount
+      requires b > 0 else InvalidAmount
+      call Token.transfer(from = 0, to = 1, amount = a, authority = 0)
+      call Token.transfer(from = 0, to = 2, amount = b, authority = 0)
+      effect { vault_balance -= a }
+      ensures state.vault_balance == old(state.vault_balance) - a - b
+    }"#;
+        let spec = crate::chumsky_adapter::parse_str(src).expect("spec parses");
+        let warnings = check_multi_cpi_same_field(&spec);
+        let hit = warnings
+            .iter()
+            .find(|w| w.rule == "multi_cpi_same_field")
+            .unwrap_or_else(|| {
+                panic!(
+                    "multi_cpi_same_field must fire; got: {:?}",
+                    warnings.iter().map(|w| &w.rule).collect::<Vec<_>>()
+                )
+            });
+        assert_eq!(hit.severity, Severity::Info);
+        assert_eq!(hit.priority, 2);
+        assert!(
+            hit.message.contains("'vault_balance'"),
+            "message must name the shared field; got: {}",
+            hit.message
+        );
+        assert!(
+            hit.message.contains("Token.transfer"),
+            "message must name the call pair; got: {}",
+            hit.message
+        );
+        assert_eq!(hit.subject.as_deref(), Some("split"));
+    }
+
+    /// Disjoint Token.transfer resources are handled by the Pinocchio
+    /// impl-targeted token projection backend. The abstract callee fields
+    /// are the same, but the generated proof reads and asserts each token
+    /// account's concrete amount independently.
+    #[test]
+    fn multi_cpi_same_field_silent_on_disjoint_token_transfer_resources() {
+        let src = r#"spec MultiCpiDisjointToken
+    program_id "11111111111111111111111111111111"
+
+    interface Token {
+      program_id "11111111111111111111111111111111"
+      handler transfer (amount : U64) {
+        accounts {
+          from      : writable
+          to        : writable
+          authority : signer
+        }
+        requires amount > 0
+        ensures state.from_balance == old(state.from_balance) - amount
+        ensures state.to_balance == old(state.to_balance) + amount
+      }
+    }
+
+    state { from_balance : U64, to_balance : U64 }
+
+    handler swap_like (a : U64) (b : U64) {
+      permissionless
+      requires a > 0 else InvalidAmount
+      requires b > 0 else InvalidAmount
+      call Token.transfer(from = user_input, to = hub_input, amount = a, authority = auth)
+      call Token.transfer(from = hub_output, to = user_output, amount = b, authority = auth)
+      ensures state.from_balance == old(state.from_balance) - a
+    }"#;
+        let spec = crate::chumsky_adapter::parse_str(src).expect("spec parses");
+        let warnings = check_multi_cpi_same_field(&spec);
+        assert!(
+            warnings.is_empty(),
+            "disjoint Token.transfer resources use per-account projections; got: {:?}",
+            warnings.iter().map(|w| &w.message).collect::<Vec<_>>()
+        );
+    }
+
+    /// Two CPI calls whose substituted ensures reference disjoint
+    /// caller-state fields → lint stays silent. No (pre, post) snapshot
+    /// pair is shared, so the over-constraint risk doesn't apply.
+    #[test]
+    fn multi_cpi_same_field_silent_on_disjoint_fields() {
+        let src = r#"spec MultiCpiDisjoint
+    program_id "11111111111111111111111111111111"
+
+    interface VaultA {
+      program_id "11111111111111111111111111111111"
+      handler debit (amount : U64) {
+        accounts { vault : writable }
+        requires amount > 0
+        ensures state.vault_a_balance == old(state.vault_a_balance) - amount
+      }
+    }
+
+    interface VaultB {
+      program_id "11111111111111111111111111111111"
+      handler debit (amount : U64) {
+        accounts { vault : writable }
+        requires amount > 0
+        ensures state.vault_b_balance == old(state.vault_b_balance) - amount
+      }
+    }
+
+    state { vault_a_balance : U64, vault_b_balance : U64 }
+
+    handler tap_both (a : U64) (b : U64) {
+      permissionless
+      requires a > 0 else InvalidAmount
+      requires b > 0 else InvalidAmount
+      call VaultA.debit(amount = a)
+      call VaultB.debit(amount = b)
+      effect { vault_a_balance -= a }
+      effect { vault_b_balance -= b }
+      ensures state.vault_a_balance == old(state.vault_a_balance) - a
+      ensures state.vault_b_balance == old(state.vault_b_balance) - b
+    }"#;
+        let spec = crate::chumsky_adapter::parse_str(src).expect("spec parses");
+        let warnings = check_multi_cpi_same_field(&spec);
+        assert!(
+            warnings.is_empty(),
+            "disjoint-field CPI ensures must not fire multi_cpi_same_field; got: {:?}",
+            warnings.iter().map(|w| &w.message).collect::<Vec<_>>()
+        );
+    }
+
+    /// Tier-0 callees (no `ensures` declared) → no substituted field
+    /// references → lint stays silent regardless of CPI multiplicity.
+    /// Catches the spec-shape where the user hasn't yet declared the
+    /// callee's contract; the `cpi_no_callee_ensures` lint surfaces
+    /// that gap separately.
+    #[test]
+    fn multi_cpi_same_field_silent_on_tier0_callees() {
+        let src = r#"spec MultiCpiTier0
+    program_id "11111111111111111111111111111111"
+
+    interface Logger {
+      program_id "11111111111111111111111111111111"
+      handler log (msg : U64) {
+        accounts { sink : writable }
+      }
+    }
+
+    state { counter : U64 }
+
+    handler tick_twice (a : U64) (b : U64) {
+      permissionless
+      requires a > 0 else InvalidAmount
+      requires b > 0 else InvalidAmount
+      call Logger.log(msg = a)
+      call Logger.log(msg = b)
+      effect { counter += a }
+      ensures state.counter == old(state.counter) + a
+    }"#;
+        let spec = crate::chumsky_adapter::parse_str(src).expect("spec parses");
+        let warnings = check_multi_cpi_same_field(&spec);
+        assert!(
+            warnings.is_empty(),
+            "tier-0 callees produce no field refs → lint must stay silent; got: {:?}",
+            warnings.iter().map(|w| &w.message).collect::<Vec<_>>()
+        );
+    }
+
+    // ──────────────────────────────────────────────────────────────────────
+    // [shape_only_cpi] lint
+    // ──────────────────────────────────────────────────────────────────────
+
+    /// Declared Tier-0 interfaces with no `ensures` must not fire
+    /// `shape_only_cpi` — firing would force `ensures true` tautologies on
+    /// handlers with no meaningful post-condition. The lint still fires for
+    /// undeclared interfaces / missing handlers (real spec bugs).
+    #[test]
+    fn shape_only_cpi_silent_on_declared_tier0_interface() {
+        let src = r#"spec Demo
+
+    interface Token {
+      program_id "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA"
+      handler transfer (amount : U64) {
+        accounts {
+          from      : writable
+          to        : writable
+          authority : signer
+        }
+      }
+    }
+
+    handler pay : State.A -> State.A {
+      call Token.transfer(from = src_ta, to = dst_ta, amount = 1)
+    }
+    "#;
+        let parsed = crate::chumsky_adapter::parse_str(src).unwrap();
+        let ws = check_completeness(&parsed);
+        let hits: Vec<_> = ws.iter().filter(|w| w.rule == "shape_only_cpi").collect();
+        assert!(
+            hits.is_empty(),
+            "Tier-0 interface with no `ensures` should not fire shape_only_cpi; got: {:?}",
+            hits.iter().map(|w| &w.message).collect::<Vec<_>>()
+        );
+    }
+
+    #[test]
+    fn shape_only_cpi_fires_on_undeclared_interface() {
+        let src = r#"spec Demo
+
+    handler pay : State.A -> State.A {
+      call Jupiter.swap(pool = amm, amount_in = 100, min_out = 90)
+    }
+    "#;
+        let parsed = crate::chumsky_adapter::parse_str(src).unwrap();
+        let ws = check_completeness(&parsed);
+        let hits: Vec<_> = ws.iter().filter(|w| w.rule == "shape_only_cpi").collect();
+        assert_eq!(
+            hits.len(),
+            1,
+            "expected one shape_only_cpi warning, got {:?}",
+            ws
+        );
+        assert!(hits[0].message.contains("not declared"));
+    }
+
+    #[test]
+    fn shape_only_cpi_silent_on_tier1_interface() {
+        // Interface declares at least one ensures — no lint should fire.
+        let src = r#"spec Demo
+
+    interface Token {
+      program_id "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA"
+      handler transfer (amount : U64) {
+        accounts {
+          from      : writable
+          to        : writable
+          authority : signer
+        }
+        ensures amount > 0
+      }
+    }
+
+    handler pay : State.A -> State.A {
+      call Token.transfer(from = src_ta, to = dst_ta, amount = 1)
+    }
+    "#;
+        let parsed = crate::chumsky_adapter::parse_str(src).unwrap();
+        let ws = check_completeness(&parsed);
+        let hits: Vec<_> = ws.iter().filter(|w| w.rule == "shape_only_cpi").collect();
+        assert!(
+            hits.is_empty(),
+            "Tier 1 interfaces should not lint, got: {:?}",
+            hits
+        );
+    }
+
+    // ----- cpi_unverified_callee P2 lint -----
+
+    #[test]
+    fn cpi_unverified_callee_fires_on_unverified_import() {
+        // Simulates an `import Token from "..."` whose provider didn't
+        // ship a proof package. The resolver wouldn't have populated
+        // `verified_callees` so the lint should fire.
+        let src = r#"spec Demo
+
+    import Token from "spl_token"
+
+    interface Token {
+      program_id "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA"
+      upstream { binary_hash "sha256:0000" }
+      handler transfer (amount : U64) {
+        accounts {
+          from      : writable
+          to        : writable
+          authority : signer
+        }
+        ensures amount > 0
+      }
+    }
+
+    handler pay : State.A -> State.A {
+      call Token.transfer(from = src_ta, to = dst_ta, amount = 1)
+    }
+    "#;
+        let parsed = crate::chumsky_adapter::parse_str(src).unwrap();
+        let ws = check_cpi_unverified_callee(&parsed);
+        assert_eq!(
+            ws.len(),
+            1,
+            "expected one unverified-callee warning; got: {ws:?}"
+        );
+        assert_eq!(ws[0].rule, "cpi_unverified_callee");
+        assert_eq!(ws[0].priority, 2);
+        assert!(ws[0].message.contains("Stance-1 axiom"));
+        assert!(ws[0].fix.contains(".qed/proofs"));
+        assert!(
+            ws[0].fix.contains("tokenProofs"),
+            "fix message should name the expected lake package; got: {}",
+            ws[0].fix
+        );
+    }
+
+    #[test]
+    fn cpi_unverified_callee_silent_when_verified_callees_lists_iface() {
+        // Same shape but `verified_callees` has the import registered,
+        // simulating a provider that did ship proofs.
+        let src = r#"spec Demo
+
+    import Token from "spl_token"
+
+    interface Token {
+      program_id "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA"
+      upstream { binary_hash "sha256:0000" }
+      handler transfer (amount : U64) {
+        accounts {
+          from      : writable
+          to        : writable
+          authority : signer
+        }
+        ensures amount > 0
+      }
+    }
+
+    handler pay : State.A -> State.A {
+      call Token.transfer(from = src_ta, to = dst_ta, amount = 1)
+    }
+    "#;
+        let mut parsed = crate::chumsky_adapter::parse_str(src).unwrap();
+        parsed
+            .verified_callees
+            .insert("Token".to_string(), std::path::PathBuf::from("/tmp/x"));
+        let ws = check_cpi_unverified_callee(&parsed);
+        assert!(
+            ws.is_empty(),
+            "verified callee should suppress the lint; got: {ws:?}"
+        );
+    }
+
+    #[test]
+    fn cpi_unverified_callee_silent_on_in_spec_interfaces() {
+        // Interface declared inline (no `import` statement) — the
+        // author owns both the contract and the call, so there's no
+        // external trust gap to surface.
+        let src = r#"spec Demo
+
+    interface Token {
+      program_id "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA"
+      upstream { binary_hash "sha256:0000" }
+      handler transfer (amount : U64) {
+        accounts {
+          from      : writable
+          to        : writable
+          authority : signer
+        }
+        ensures amount > 0
+      }
+    }
+
+    handler pay : State.A -> State.A {
+      call Token.transfer(from = src_ta, to = dst_ta, amount = 1)
+    }
+    "#;
+        let parsed = crate::chumsky_adapter::parse_str(src).unwrap();
+        let ws = check_cpi_unverified_callee(&parsed);
+        assert!(
+            ws.is_empty(),
+            "inline interface (no import) should not fire; got: {ws:?}"
+        );
+    }
+
+    #[test]
+    fn cpi_unverified_callee_silent_on_tier0_imports() {
+        // Imported interface with no `ensures` — cpi_no_callee_ensures
+        // (P1) owns that case; cpi_unverified_callee should stay quiet.
+        let src = r#"spec Demo
+
+    import Token from "spl_token"
+
+    interface Token {
+      program_id "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA"
+      handler transfer (amount : U64) {
+        accounts {
+          from      : writable
+          to        : writable
+          authority : signer
+        }
+      }
+    }
+
+    handler pay : State.A -> State.A {
+      call Token.transfer(from = src_ta, to = dst_ta, amount = 1)
+    }
+    "#;
+        let parsed = crate::chumsky_adapter::parse_str(src).unwrap();
+        let ws = check_cpi_unverified_callee(&parsed);
+        assert!(
+            ws.is_empty(),
+            "Tier-0 imports should not double-fire; got: {ws:?}"
+        );
+    }
+
+    #[test]
+    fn cpi_unverified_callee_deduplicates_repeated_calls() {
+        // Two handlers both calling Token.transfer — the lint should
+        // surface the trust-gap once per (interface, handler), not per
+        // call site.
+        let src = r#"spec Demo
+
+    import Token from "spl_token"
+
+    interface Token {
+      program_id "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA"
+      upstream { binary_hash "sha256:0000" }
+      handler transfer (amount : U64) {
+        accounts {
+          from      : writable
+          to        : writable
+          authority : signer
+        }
+        ensures amount > 0
+      }
+    }
+
+    handler pay_a : State.A -> State.A {
+      call Token.transfer(from = src_ta, to = dst_ta, amount = 1)
+    }
+
+    handler pay_b : State.A -> State.A {
+      call Token.transfer(from = src_ta, to = dst_ta, amount = 2)
+    }
+    "#;
+        let parsed = crate::chumsky_adapter::parse_str(src).unwrap();
+        let ws = check_cpi_unverified_callee(&parsed);
+        assert_eq!(ws.len(), 1, "should dedupe across call sites; got: {ws:?}");
+    }
+
+    // ------------------------------------------------------------------
+    // collect_require_verified_findings
+    // ------------------------------------------------------------------
+
+    #[test]
+    fn require_verified_fires_on_unverified_import_with_ensures() {
+        // Non-sentinel binary_hash so the sentinel exemption doesn't
+        // intercept. `verified_callees` is empty → provider shipped no
+        // proof package → finding.
+        let src = r#"spec Demo
+
+    import Token from "amm_lib"
+
+    interface Token {
+      program_id "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA"
+      upstream { binary_hash "sha256:abc123" }
+      handler transfer (amount : U64) {
+        accounts {
+          from      : writable
+          to        : writable
+          authority : signer
+        }
+        ensures amount > 0
+      }
+    }
+
+    handler pay : State.A -> State.A {
+      call Token.transfer(from = src_ta, to = dst_ta, amount = 1)
+    }
+    "#;
+        let parsed = crate::chumsky_adapter::parse_str(src).unwrap();
+        let findings = collect_require_verified_findings(&parsed);
+        assert_eq!(
+            findings.len(),
+            1,
+            "expected one finding for unverified Token; got: {findings:?}"
+        );
+        assert_eq!(findings[0].interface_name, "Token");
+        assert!(
+            findings[0].fix_hint.contains(".qed/proofs"),
+            "fix hint should point at the proof-package path; got: {}",
+            findings[0].fix_hint
+        );
+    }
+
+    #[test]
+    fn require_verified_silent_when_provider_shipped_proofs() {
+        // verified_callees populated → provider has proofs → no finding.
+        let src = r#"spec Demo
+
+    import Token from "amm_lib"
+
+    interface Token {
+      program_id "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA"
+      upstream { binary_hash "sha256:abc123" }
+      handler transfer (amount : U64) {
+        accounts {
+          from      : writable
+          to        : writable
+          authority : signer
+        }
+        ensures amount > 0
+      }
+    }
+
+    handler pay : State.A -> State.A {
+      call Token.transfer(from = src_ta, to = dst_ta, amount = 1)
+    }
+    "#;
+        let mut parsed = crate::chumsky_adapter::parse_str(src).unwrap();
+        parsed
+            .verified_callees
+            .insert("Token".to_string(), std::path::PathBuf::from("/tmp/x"));
+        let findings = collect_require_verified_findings(&parsed);
+        assert!(
+            findings.is_empty(),
+            "verified callee must suppress the finding; got: {findings:?}"
+        );
+    }
+
+    #[test]
+    fn require_verified_silent_on_tier0_imports() {
+        // No ensures clauses on any handler → Tier 0. Owned by the
+        // cpi_no_callee_ensures P1 lint, not by --require-verified.
+        let src = r#"spec Demo
+
+    import Token from "amm_lib"
+
+    interface Token {
+      program_id "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA"
+      upstream { binary_hash "sha256:abc123" }
+      handler transfer (amount : U64) {
+        accounts {
+          from      : writable
+          to        : writable
+          authority : signer
+        }
+      }
+    }
+
+    handler pay : State.A -> State.A {
+      call Token.transfer(from = src_ta, to = dst_ta, amount = 1)
+    }
+    "#;
+        let parsed = crate::chumsky_adapter::parse_str(src).unwrap();
+        let findings = collect_require_verified_findings(&parsed);
+        assert!(
+            findings.is_empty(),
+            "Tier-0 (no ensures) imports must not fire --require-verified; got: {findings:?}"
+        );
+    }
+
+    #[test]
+    fn require_verified_silent_on_sentinel_pinned_natives() {
+        // Sentinel binary_hash (sha256:00…00) marks a native program
+        // (System Program style) — the validator runtime is the trust
+        // boundary, not a proof package. `--require-verified` exempts
+        // these so any spec that imports `from "system"` doesn't
+        // false-fail.
+        let src = r#"spec Demo
+
+    import System from "system_lib"
+
+    interface System {
+      program_id "11111111111111111111111111111111"
+      upstream { binary_hash "sha256:0000000000000000000000000000000000000000000000000000000000000000" }
+      handler transfer (amount : U64) {
+        accounts {
+          from      : writable
+          to        : writable
+          authority : signer
+        }
+        ensures amount > 0
+      }
+    }
+
+    handler pay : State.A -> State.A {
+      call System.transfer(from = src_ta, to = dst_ta, amount = 1)
+    }
+    "#;
+        let parsed = crate::chumsky_adapter::parse_str(src).unwrap();
+        let findings = collect_require_verified_findings(&parsed);
+        assert!(
+            findings.is_empty(),
+            "sentinel-pinned native must be exempt; got: {findings:?}"
+        );
+    }
+
+    #[test]
+    fn require_verified_silent_on_inline_interfaces() {
+        // Interface declared inline (no `import` statement) — author
+        // owns both sides of the contract. `--require-verified` only
+        // gates on imported interfaces.
+        let src = r#"spec Demo
+
+    interface Token {
+      program_id "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA"
+      upstream { binary_hash "sha256:abc123" }
+      handler transfer (amount : U64) {
+        accounts {
+          from      : writable
+          to        : writable
+          authority : signer
+        }
+        ensures amount > 0
+      }
+    }
+
+    handler pay : State.A -> State.A {
+      call Token.transfer(from = src_ta, to = dst_ta, amount = 1)
+    }
+    "#;
+        let parsed = crate::chumsky_adapter::parse_str(src).unwrap();
+        let findings = collect_require_verified_findings(&parsed);
+        assert!(
+            findings.is_empty(),
+            "inline interfaces must not fire; got: {findings:?}"
+        );
+    }
+}
